requirements-deploy.txt
certifi @ file:///opt/conda/conda-bld/certifi_1655968806487/work/certifi
click==8.1.3
dlib==19.24.0
Flask==2.1.3
Flask-Cors==3.0.10
imageio==2.19.5
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
llvmlite==0.38.1
MarkupSafe==2.1.1
networkx==2.8.5
numba==0.55.2
numpy==1.22.4
opencv-python==4.5.1.48
packaging==21.3
Pillow==9.2.0
pyparsing==3.0.9
PyWavelets==1.3.0
scikit-image==0.19.3
scipy==1.8.1
six==1.16.0
tifffile==2022.5.4
Werkzeug==2.1.2
zipp==3.8.1
